Allow caller to define minimal number of new items.
This features a new endpoint that takes a json object {count: x, return_url: url}.
The route handler will start threads for each known imgur topic and query the corresponding endpoints
and returns.
The intermediate results are cached (with ids) and once the minimum number is reached, they will be
posted back to the caller.

Issues:
1) The threading stuff is a bit hacky (especially ensuring the callback is only called once). Need to find better solution, or scrap it.
2) Depending on the number of topics and the amount of new posts (not albums) per page, this approach can return a lot more elements, than required.
3) currently post_ids are only logged to the console. This has to be changed to load corresponding posts from mongo and post them to the return_url.