Test recommender service with surprise algorithms. Does not support retraining or unkown users (yet)
The data accessor is currently not using a database, but the interface should be reuseable for that case.
Instead the Movielens 100k dataset is currently used.
Training this set currently takes a few seconds.

Recommendation times take ~3 seconds, while being parallelized with multiple processes.
(4 cores).
Part of the time taken is the access of the n top rateditems.
Caching and or stored procedures could improve these times.